Fix TokenFileImportNotSupported in release builds
rust-embed falls back to reading from the file system in debug builds,
which allowed a bad path to be used to access embedded token files in
our tests. This fixes the path used to access the embedded token files
and enables the debug-embed feature to force rust-embed to always embed
the token files.